+++ base/src/test/java/io/github/sds100/keymapper/base/actions/PerformActionsUseCaseTest.kt
@@ -14,7 +14,6 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.junit.MockitoJUnitRunner
 import org.mockito.kotlin.any
-import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.never
@@ -100,6 +99,6 @@ class PerformActionsUseCaseTest {
             useCase.perform(action)
 
             // THEN
-            verify(mockToastAdapter, never()).show(anyOrNull())
+            verify(mockToastAdapter, never()).show(any(), any())
         }
 }