+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - #2045 add action to input on-screen keyboard enter/send button.
 - #2106 disable the keyboard auto-switching setting when manually switching the keyboard in the Key Mapper homescreen menu.
+- #1029 add action to show a toast message.
 - #2081 add getevent debug screen.
 - #2087 small segmented button text is not readable in dark mode.
 - #2077 rename "This device" and "any device" to "This Android device" and "Any input device" to prevent confusion.